# Auth/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  /auth/:
    post:
      tags:
        - auth
      summary: Authenticate
      operationId: authenticate_auth__post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_authenticate_auth__post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/register/:
    post:
      tags:
        - auth
      summary: Register User
      operationId: register_user_auth_register__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginSchema'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/change_user_data/:
    put:
      tags:
        - auth
      summary: Change User Data
      operationId: change_user_data_auth_change_user_data__put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDataSchema'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OAuth2PasswordBearer: []
  /auth/create/:
    post:
      tags:
        - tasks
      summary: Create Task
      operationId: create_task_auth_create__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskSchema'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskIdSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OAuth2PasswordBearer: []
  /auth/update/{task_id}:
    put:
      tags:
        - tasks
      summary: Update Task
      operationId: update_task_auth_update__task_id__put
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskSchema'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullTaskSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/delete/{task_id}:
    delete:
      tags:
        - tasks
      summary: Delete Task
      operationId: delete_task_auth_delete__task_id__delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteTaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/get/{task_id}:
    get:
      tags:
        - tasks
      summary: Get Task By Id
      operationId: get_task_by_id_auth_get__task_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullTaskSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/get_all/:
    get:
      tags:
        - tasks
      summary: Get Tasks
      operationId: get_tasks_auth_get_all__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: page_number
          in: query
          required: true
          schema:
            type: integer
            exclusiveMinimum: 0
            title: Page Number
        - name: page_size
          in: query
          required: true
          schema:
            type: integer
            exclusiveMinimum: 0
            title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FullTaskSchema'
                title: Response Get Tasks Auth Get All  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_authenticate_auth__post:
      properties:
        grant_type:
          anyOf:
            - type: string
              pattern: password
            - type: 'null'
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
        scope:
          type: string
          title: Scope
          default: ''
        client_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
            - type: string
            - type: 'null'
          title: Client Secret
      type: object
      required:
        - username
        - password
      title: Body_authenticate_auth__post
    DeleteTaskResponse:
      properties:
        success:
          type: boolean
          title: Success
      type: object
      required:
        - success
      title: DeleteTaskResponse
    FullTaskSchema:
      properties:
        task_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Task Id
        title:
          type: string
          title: Title
        text:
          type: string
          title: Text
      type: object
      required:
        - task_id
        - title
        - text
      title: FullTaskSchema
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    TaskIdSchema:
      properties:
        task_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Task Id
      type: object
      required:
        - task_id
      title: TaskIdSchema
    TaskSchema:
      properties:
        title:
          type: string
          title: Title
        text:
          type: string
          title: Text
      type: object
      required:
        - title
        - text
      title: TaskSchema
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
      type: object
      required:
        - access_token
        - token_type
      title: Token
    UserDataSchema:
      properties:
        first_name:
          anyOf:
            - type: string
            - type: 'null'
          title: First Name
        last_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Last Name
        birth_date:
          anyOf:
            - type: string
              format: date
            - type: 'null'
          title: Birth Date
        email:
          anyOf:
            - type: string
              format: email
            - type: 'null'
          title: Email
        phone_number:
          anyOf:
            - type: string
            - type: 'null'
          title: Phone Number
      type: object
      required:
        - first_name
        - last_name
        - birth_date
        - email
        - phone_number
      title: UserDataSchema
    UserLoginSchema:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
      type: object
      required:
        - username
        - password
      title: UserLoginSchema
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth